+++ packages/core/navigation-next/src/renderer/components.js
@@ -13,9 +13,11 @@ import ShipIcon from '@atlaskit/icon/glyph/ship';
 import { gridSize as gridSizeFn } from '@atlaskit/theme';
 
 import {
+  ContainerHeader,
+  ContainerViewSubscriber,
   Item as BaseItem,
   ItemPrimitive,
-  NavAPISubscriber,
+  RootViewSubscriber,
   Section,
   SectionSeparator,
   SectionTitle,
@@ -55,10 +57,13 @@ const GoToItem = ({ after: afterProp, goTo, ...rest }: GoToItemProps) => {
   }
 
   const props = { ...rest, after };
+  const ViewSubscriber = goTo.match(/^root\//)
+    ? RootViewSubscriber
+    : ContainerViewSubscriber;
   return (
-    <NavAPISubscriber>
-      {api => <Item onClick={() => api.setView(goTo)} {...props} />}
-    </NavAPISubscriber>
+    <ViewSubscriber>
+      {view => <Item onClick={() => view.setView(goTo)} {...props} />}
+    </ViewSubscriber>
   );
 };
 
@@ -168,6 +173,7 @@ const Nested = ({
 );
 
 const itemComponents = {
+  ContainerHeader,
   Debug,
   GoToItem,
   Item,