+++ packages/editor/editor-core/__tests__/plugins/media/index.ts
@@ -71,7 +71,7 @@ describe('Media plugin', () => {
           allowMediaSingle: true,
           customDropzoneContainer: dropzoneContainer,
         }),
-        codeBlockPlugin,
+        codeBlockPlugin(),
         rulePlugin,
         tablePlugin,
       ],