+++ packages/editor/editor-core/src/editor.tsx
@@ -9,6 +9,7 @@ import { EditorProps } from './types';
 import { ReactEditorView } from './create-editor';
 import { EventDispatcher } from './event-dispatcher';
 import EditorContext from './ui/EditorContext';
+import { PortalProvider, PortalRenderer } from './ui/PortalProvider';
 
 export * from './types';
 
@@ -64,7 +65,9 @@ export default class Editor extends React.Component<EditorProps, {}> {
     );
     if (this.props.shouldFocus) {
       if (!instance.view.hasFocus()) {
-        instance.view.focus();
+        setTimeout(() => {
+          instance.view.focus();
+        }, 0);
       }
     }
   }
@@ -182,38 +185,48 @@ export default class Editor extends React.Component<EditorProps, {}> {
 
     return (
       <EditorContext editorActions={this.editorActions}>
-        <ReactEditorView
-          editorProps={overriddenEditorProps}
-          providerFactory={this.providerFactory}
-          onEditorCreated={this.onEditorCreated}
-          onEditorDestroyed={this.onEditorDestroyed}
-          render={({ editor, view, eventDispatcher, config }) => (
-            <Component
-              disabled={this.props.disabled}
-              editorActions={this.editorActions}
-              editorDOMElement={editor}
-              editorView={view}
-              providerFactory={this.providerFactory}
-              eventDispatcher={eventDispatcher}
-              maxHeight={this.props.maxHeight}
-              onSave={this.props.onSave ? this.handleSave : undefined}
-              onCancel={this.props.onCancel}
-              popupsMountPoint={this.props.popupsMountPoint}
-              popupsBoundariesElement={this.props.popupsBoundariesElement}
-              contentComponents={config.contentComponents}
-              primaryToolbarComponents={config.primaryToolbarComponents}
-              secondaryToolbarComponents={config.secondaryToolbarComponents}
-              insertMenuItems={this.props.insertMenuItems}
-              customContentComponents={this.props.contentComponents}
-              customPrimaryToolbarComponents={
-                this.props.primaryToolbarComponents
-              }
-              customSecondaryToolbarComponents={
-                this.props.secondaryToolbarComponents
-              }
-              addonToolbarComponents={this.props.addonToolbarComponents}
-              collabEdit={this.props.collabEdit}
-            />
+        <PortalProvider
+          render={portalProviderAPI => (
+            <>
+              <ReactEditorView
+                editorProps={overriddenEditorProps}
+                portalProviderAPI={portalProviderAPI}
+                providerFactory={this.providerFactory}
+                onEditorCreated={this.onEditorCreated}
+                onEditorDestroyed={this.onEditorDestroyed}
+                render={({ editor, view, eventDispatcher, config }) => (
+                  <Component
+                    disabled={this.props.disabled}
+                    editorActions={this.editorActions}
+                    editorDOMElement={editor}
+                    editorView={view}
+                    providerFactory={this.providerFactory}
+                    eventDispatcher={eventDispatcher}
+                    maxHeight={this.props.maxHeight}
+                    onSave={this.props.onSave ? this.handleSave : undefined}
+                    onCancel={this.props.onCancel}
+                    popupsMountPoint={this.props.popupsMountPoint}
+                    popupsBoundariesElement={this.props.popupsBoundariesElement}
+                    contentComponents={config.contentComponents}
+                    primaryToolbarComponents={config.primaryToolbarComponents}
+                    secondaryToolbarComponents={
+                      config.secondaryToolbarComponents
+                    }
+                    insertMenuItems={this.props.insertMenuItems}
+                    customContentComponents={this.props.contentComponents}
+                    customPrimaryToolbarComponents={
+                      this.props.primaryToolbarComponents
+                    }
+                    customSecondaryToolbarComponents={
+                      this.props.secondaryToolbarComponents
+                    }
+                    addonToolbarComponents={this.props.addonToolbarComponents}
+                    collabEdit={this.props.collabEdit}
+                  />
+                )}
+              />
+              <PortalRenderer portalProviderAPI={portalProviderAPI} />
+            </>
           )}
         />
       </EditorContext>