+++ packages/editor/editor-core/src/plugins/text-formatting/pm-plugins/input-rule.ts
@@ -123,7 +123,7 @@ function addCodeMark(
     }
     analyticsService.trackEvent('atlassian.editor.format.code.autoformatting');
     const regexStart = end - match[2].length + 1;
-    return transformToCodeAction(state, regexStart, end)
+    return transformToCodeAction(regexStart, end, state.tr)
       .delete(regexStart, regexStart + specialChar.length)
       .removeStoredMark(markType);
   };