+++ packages/media/media-test-helpers/src/fakeContextProvider.ts
@@ -1,6 +1,5 @@
 import { Observable } from 'rxjs';
-import { Context } from '@atlaskit/media-core';
-import { ContextConfig } from '../../media-store';
+import { Context, ContextConfig } from '@atlaskit/media-core';
 
 const defaultContextConfig = {
   serviceHost: 'some-service-host',
@@ -32,6 +31,7 @@ export const fakeContext = (
   const getUrlPreviewProvider = returns({
     observable: returns(Observable.of('nothing')),
   });
+  const getFile = jest.fn();
   const getLocalPreview = jest.fn();
   const setLocalPreview = jest.fn();
   const removeLocalPreview = jest.fn();
@@ -39,6 +39,7 @@ export const fakeContext = (
   const getBlobService = jest.fn();
   const uploadFile = jest.fn();
   const defaultContext: Context = {
+    getFile,
     getBlobService,
     getLocalPreview,
     setLocalPreview,