+++ packages/search/global-search/src/api/PeopleSearchClient.ts
@@ -36,6 +36,8 @@ export default class PeopleSearchClientImpl implements PeopleSearchClient {
   private serviceConfig: ServiceConfig;
   private cloudId: string;
 
+  private readonly RESULT_LIMIT = 5;
+
   constructor(url: string, cloudId: string) {
     this.serviceConfig = { url: url };
     this.cloudId = cloudId;
@@ -87,7 +89,7 @@ export default class PeopleSearchClientImpl implements PeopleSearchClient {
       variables: {
         cloudId: this.cloudId,
         displayName: query,
-        first: 5,
+        first: this.RESULT_LIMIT,
         offset: 1,
         excludeInactive: true,
         excludeBots: true,