+++ packages/search/global-search/src/components/confluence/ConfluenceSearchResults.tsx
@@ -94,7 +94,7 @@ export default function searchResults(props: Props) {
     return [
       renderObjectsGroup(
         'Recent pages and blogs',
-        take(recentlyViewedPages, 7),
+        take(recentlyViewedPages, 8),
         query,
       ),
       renderSpacesGroup('Recent spaces', take(recentlyViewedSpaces, 3), query),
@@ -109,10 +109,10 @@ export default function searchResults(props: Props) {
   return [
     renderObjectsGroup(
       'Pages, blogs and attachments',
-      take(objectResults, 5),
+      take(objectResults, 8),
       query,
     ),
-    renderSpacesGroup('Spaces', take(spaceResults, 5), query),
+    renderSpacesGroup('Spaces', take(spaceResults, 3), query),
     renderPeopleGroup('People', take(peopleResults, 3), query),
   ];
 }